company/ listing add new and delete functionality implemented.
notes/ listings management is made mainly on SKUs. creating a new listing, wiil only ask for SKU number. Duplicate prevention validation is made in server. Only SKU number will be entered in DB initially; the rest of the fields contain default values. Same as for deleting listing entries. It is made mainly with the SKU number and server will pull out from the array in DB. All of this is made in PUT route.

additional/ started POST route for new products. It will create the main document, with no listing entries. Only when the main document is created, then the user will "edit" the main document to create listing entries